- Update readme and refactor code
