+++ src/main/kotlin/com/piashcse/di/KoinModule.kt
@@ -1,22 +1,21 @@
 package com.piashcse.di
 
-import com.piashcse.controller.*
-import com.piashcse.modules.auth.service.AuthController
-import com.piashcse.modules.brand.service.BrandController
-import com.piashcse.modules.cart.service.CartController
-import com.piashcse.modules.consent.service.ConsentController
-import com.piashcse.modules.order.service.OrderController
-import com.piashcse.modules.payment.service.PaymentController
-import com.piashcse.modules.policy.service.PolicyController
-import com.piashcse.modules.product.service.ProductController
-import com.piashcse.modules.productcategory.service.ProductCategoryController
-import com.piashcse.modules.productsubcategory.service.ProductSubCategoryController
-import com.piashcse.modules.profile.service.ProfileController
-import com.piashcse.modules.review_rating.service.ReviewRatingController
-import com.piashcse.modules.shipping.service.ShippingController
-import com.piashcse.modules.shop.service.ShopController
-import com.piashcse.modules.shopcategory.service.ShopCategoryController
-import com.piashcse.modules.wishlist.service.WishListController
+import com.piashcse.modules.auth.controller.AuthController
+import com.piashcse.modules.brand.controller.BrandController
+import com.piashcse.modules.cart.controller.CartController
+import com.piashcse.modules.consent.controller.ConsentController
+import com.piashcse.modules.order.controller.OrderController
+import com.piashcse.modules.payment.controller.PaymentController
+import com.piashcse.modules.policy.controller.PolicyController
+import com.piashcse.modules.product.controller.ProductController
+import com.piashcse.modules.productcategory.controller.ProductCategoryController
+import com.piashcse.modules.productsubcategory.controller.ProductSubCategoryController
+import com.piashcse.modules.profile.controller.ProfileController
+import com.piashcse.modules.review_rating.controller.ReviewRatingController
+import com.piashcse.modules.shipping.controller.ShippingController
+import com.piashcse.modules.shop.controller.ShopController
+import com.piashcse.modules.shopcategory.controller.ShopCategoryController
+import com.piashcse.modules.wishlist.controller.WishListController
 import org.koin.dsl.module
 
 val controllerModule = module {
