+++ src/main/kotlin/com/piashcse/plugins/ConfigureRouting.kt
@@ -1,38 +1,37 @@
 package com.piashcse.plugins
 
-import com.piashcse.controller.*
 import com.piashcse.modules.auth.routes.authRoutes
-import com.piashcse.modules.auth.service.AuthController
+import com.piashcse.modules.auth.controller.AuthController
 import com.piashcse.modules.brand.routes.brandRoutes
-import com.piashcse.modules.brand.service.BrandController
+import com.piashcse.modules.brand.controller.BrandController
 import com.piashcse.modules.cart.routes.cartRoutes
-import com.piashcse.modules.cart.service.CartController
+import com.piashcse.modules.cart.controller.CartController
 import com.piashcse.modules.consent.routes.consentRoutes
-import com.piashcse.modules.consent.service.ConsentController
+import com.piashcse.modules.consent.controller.ConsentController
 import com.piashcse.modules.order.routes.orderRoutes
-import com.piashcse.modules.order.service.OrderController
+import com.piashcse.modules.order.controller.OrderController
 import com.piashcse.modules.payment.routes.paymentRoutes
-import com.piashcse.modules.payment.service.PaymentController
+import com.piashcse.modules.payment.controller.PaymentController
 import com.piashcse.modules.policy.routes.policyRoutes
-import com.piashcse.modules.policy.service.PolicyController
+import com.piashcse.modules.policy.controller.PolicyController
 import com.piashcse.modules.product.routes.productRoutes
-import com.piashcse.modules.product.service.ProductController
+import com.piashcse.modules.product.controller.ProductController
 import com.piashcse.modules.productcategory.routes.productCategoryRoutes
-import com.piashcse.modules.productcategory.service.ProductCategoryController
+import com.piashcse.modules.productcategory.controller.ProductCategoryController
 import com.piashcse.modules.productsubcategory.routes.productSubCategoryRoutes
-import com.piashcse.modules.productsubcategory.service.ProductSubCategoryController
+import com.piashcse.modules.productsubcategory.controller.ProductSubCategoryController
 import com.piashcse.modules.profile.routes.profileRoutes
-import com.piashcse.modules.profile.service.ProfileController
+import com.piashcse.modules.profile.controller.ProfileController
 import com.piashcse.modules.review_rating.routes.reviewRatingRoutes
-import com.piashcse.modules.review_rating.service.ReviewRatingController
+import com.piashcse.modules.review_rating.controller.ReviewRatingController
 import com.piashcse.modules.shipping.routes.shippingRoutes
-import com.piashcse.modules.shipping.service.ShippingController
-import com.piashcse.modules.shop.service.ShopController
+import com.piashcse.modules.shipping.controller.ShippingController
+import com.piashcse.modules.shop.controller.ShopController
 import com.piashcse.modules.shop.routes.shopRoutes
 import com.piashcse.modules.shopcategory.routes.shopCategoryRoutes
-import com.piashcse.modules.shopcategory.service.ShopCategoryController
+import com.piashcse.modules.shopcategory.controller.ShopCategoryController
 import com.piashcse.modules.wishlist.routes.wishListRoutes
-import com.piashcse.modules.wishlist.service.WishListController
+import com.piashcse.modules.wishlist.controller.WishListController
 import io.ktor.server.application.*
 import io.ktor.server.routing.*
 import org.koin.ktor.ext.inject
