+++ src/main/resources/hikari.properties
@@ -1,6 +1,6 @@
 dataSourceClassName=org.postgresql.ds.PGSimpleDataSource
 dataSource.user=postgres
 dataSource.password=p123
-dataSource.databaseName=ktor-1.0.0
+dataSource.databaseName=ktor-1.0.3
 dataSource.portNumber=5432
 dataSource.serverName=localhost
